fix: extract_colors 병렬 처리 + color_hex 추출 완료 반영
- color_extractor.py: float64→float32 (K-means RuntimeWarning 수정)
- extract_colors.py: ThreadPoolExecutor 배치 병렬 처리 + 중간 저장
- TASK.md: Task 2.12 color_hex 미추출 경고 → 완료 (PCF 82.9, CH 81.3)

+++ backend/app/services/color_extractor.py
@@ -54,7 +54,7 @@ def extract_colors_from_image(
     img = img.convert("RGB")
     img = _resize_for_clustering(img)
 
-    pixels = np.array(img).reshape(-1, 3).astype(np.float64)
+    pixels = np.array(img).reshape(-1, 3).astype(np.float32)
     pixels = _remove_background(pixels)
 
     if len(pixels) < n_colors:
